// mobile/src/models/sub_summary.model.ts
import z from "zod";

export const SubSummarySchema = z.object({
  id: z.string(),
  summary_id: z.string(),
  name: z.string(),
  total: z.string(),
});

// insertion schema (id generated by SQLite / ULID)
export const CreateSubSummarySchema = SubSummarySchema.omit({
  id: true,
});

export type SubSummary = z.infer<typeof SubSummarySchema>;

export type CreateSubSummary = z.infer<typeof CreateSubSummarySchema>;